Started Retail-Inventory Project
We have started an retail inventory project in which there are 3 different types of users:- superadmin, admin and executive. We have done the work of superadmin in this and maximum work admin in which he can add their products after adding categories in this project. For that, we have used mixed schema provided by mongoose and they can also manage the history of rate and purchase by entering privacy password.